move parameter store into services directory
