+++ lib/makeOprah.js
@@ -4,7 +4,7 @@ const chalk = require('chalk');
 const { log } = require('../lib/utils/logger');
 const { getConfig } = require('../lib/utils/config');
 const { populateConfig, populateSecret } = require('./services');
-const { makeParameterStore } = require('./make-parameter-store');
+const { makeParameterStore } = require('./services/parameter-store/make-parameter-store');
 const { makeList } = require('./services/list-config');
 const { makeInit } = require('./services/make-init');
