+++ lib/services/parameter-store/make-parameter-store.js
@@ -1,12 +1,11 @@
 'use strict';
 
-const Bluebird = require('bluebird');
 const get = require('lodash.get');
 const ssmStore = require('./stores/ssm');
 const ddbStore = require('./stores/ddb');
 
 const chalk = require('chalk');
-const { log } = require('./utils/logger');
+const { log } = require('../../utils/logger');
 
 const makeParameterStore = ({ getSettings }) => {
     const getProvider = () => {
